+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - PixInternalTransactionReport resource
 - PixInternalTransactionReport.Log resource
 - priority and reason attributes to PixRequest resource
+- rules and debtorWorkspaceId parameters to CreditNote resource
 
 ## [0.16.0] - 2026-06-01
 ### Added